Increase Drankdozijn robustness
Before it was a swing and a miss whether the drankdozijn scraper would be able to fully
collect all whiskys from the website. This MR changes this by scrolling to the last visible
element and using a page down action, rather than attempting to scroll to the end of the
page regardless of the current position of the view.

Additionally, this allows the chrome driver to run headless. Thus, making the scraping
process for drankdozijn a lot more robust